Capture/store jaw opened/closed values. Enable the editor capture buttons only at runtime

+++ Assets/_ScriptsNew/MrPuppetHubConnection.cs
@@ -40,7 +40,7 @@ namespace Thinko.MrPuppet
         public Quaternion ShoulderRotation;
 
         [ReadOnly]
-        public float Jaw;
+        public int Jaw;
 
         [Header("Debug")]
         public bool OutputData = false;
